fix: 🐛 Code review. Changed:Folder organization and dicom file
Move fileLoaderService and others to a specific folder. When loading
dicom file change to only retrieve the file(not use cornerstone to cache
or anything else).
