+++ platform/viewer/src/lib/fileLoaders/fileLoader.js
@@ -1,0 +1,6 @@
+export default class FileLoader {
+  fileType;
+  loadFile(file, imageId) { }
+  getDataset(image, imageId) { }
+  getStudies(dataset, imageId) { }
+}
